Add script to merge a new .pot template into an existing po translation
This script updates msgids in a po file with msgids found in a
template where the msgctxt is the same in both.  Because Mahara should
change the string key (msgctxt) whenever a string changes its meaning,
this script can be used to avoid the problem where small typo fixes in
English strings result in previously translated strings appearing
untranslated in launchpad.  It should be run over existing po files
whenever a new template is created, and the resulting po output file
pushed back up to launchpad along with the new template.