When pushing new .pot templates to launchpad, update .po files too
This change modifies the launchpad .pot updater.  After checking the
English strings for changes and committing the new .pot template, it
then pulls the latest exported po files down from launchpad, updates
any msgids in them to reflect the changes in the newly created
template, and pushes the resulting .po files to launchpad along with
the new .pot.

This won't help until the launchpad settings are modified to import
.po translations from the bzr branch as well as templates.