Fix the check for updated msgids
When updating the msgids in an existing po translation, the diff from
the existing .po file to the the update-po-from-pot.pl output was only
checking lines starting with "[+-]msg", in order to avoid pushing the
file every time there was a change to the .po header.

This needs to be changed because diffs in help files (and probably to
the second and later lines of multi-line translations) don't start
with 'msgid'.

Instead, attempt to find diffs on lines following the "X-Generator:
Launchpad" or "X-Launchpad-Export-Date:" lines (hopefully in the
header).