change-merged: Removing leading whitespace from each line
In order to have compatibility with the launchpad email API.
Leading whitespace followed by a command word is how you send
a command; which means any commit message that has a line starting
with the word "bug" is interpreted as a command and errors out.
The easiest fix is to strip the laeding whitespace. (An alternative
fix would be to preface leading whitespace with a punctuation mark.)