Patching lptools to send the auto-generated release files
We were *always* getting prompted for the changes & releasenotes
three times on each release. That's because the optional lptools
arguments for change & release notes must be the 5th and 6th.
But, we couldn't provide them, because the lptools 4th optional
argument is "new milestone", and if you try to provide one that
already exists, it errors out.

So, we had to leave "new milestone" empty, and that meant we
also had to leave change & release notes empty.

Rather than getting deep into the logic of the lptools utility,
I just patched is so that it always skips the "new milestone"
argument.