Prevent automatic running of code submitted by untrusted Gerrit users
For security reasons, it's unwise to allow running of arbitrary code
uploaded my members of the general public. This test should limit
automated test running, to code uploaded by users who have been
manually added to the "Mahara Testers" or "Mahara Reviewers" groups.
(And/or code that has a +2 code review)

The check is against the code *uploader*, not the Gerrit change *owner*,
so if a patch is rebased by a trusted user, it will trigger automatic
code execution.